We got Playback!
Added code to start and stop video playback, as well as preliminary
styling for the player. Note that .MKV videos with DTS or Dolby audio
codes don't play sound because said audio formats are not supported by
the standard HTML5 player. Maybe we should look into transcoding or a
different player?